Refactor the test process of bench2.c
Originally, bench2 will run QF and CQF in both memory and disk,
but since the difference between in-memory and in-disk operation is
the initialize and close function, the code is highly duplicated.

To reduce duplication, I add command option for it, now it will only
run in-memory or in-disk depending on the command option.
( -m for mem, -d for disk )

I also change the performance measurment from simple timing to
calulating Million/Thousand Operation per Second each 5% load factor,
and store the result in files, so we can do further visualizing.