Fix bug that will cause undefined behaviour
According to C99 standard 6.5.7.3, if right operand of shift operation
is greater than or equal to the width of the promoted left operand,
the behavior is undefined.
In short, we can't shift more than 63 bits in this case.

To fix this, we have to add another checking process about the
bit length of q + r.